requirements.txt
<<<<<<< HEAD
-i https://pypi.org/simple
aiohappyeyeballs==2.4.0; python_version >= '3.8'
aiohttp==3.10.5; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
attrs==24.2.0; python_version >= '3.7'
black==24.8.0; python_version >= '3.8'
certifi==2024.7.4; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
distro==1.9.0; python_version >= '3.6'
filelock==3.15.4; python_version >= '3.8'
frozenlist==1.4.1; python_version >= '3.8'
fsspec==2024.6.1; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
huggingface-hub==0.24.6; python_full_version >= '3.8.0'
idna==3.7; python_version >= '3.5'
jiter==0.5.0; python_version >= '3.8'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
langchain==0.2.14; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-community==0.2.12; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-core==0.2.34; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-ollama==0.1.1; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-openai==0.1.22; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-pinecone==0.1.3; python_version < '3.13' and python_full_version >= '3.8.1'
langchain-text-splitters==0.2.2; python_version < '4.0' and python_full_version >= '3.8.1'
langsmith==0.1.101; python_version < '4.0' and python_full_version >= '3.8.1'
marshmallow==3.22.0; python_version >= '3.8'
multidict==6.0.5; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
numpy==1.26.4; python_version >= '3.12'
ollama==0.3.1; python_version >= '3.8' and python_version < '4.0'
openai==1.42.0; python_full_version >= '3.7.1'
orjson==3.10.7; python_version >= '3.8'
packaging==24.1; python_version >= '3.8'
pathspec==0.12.1; python_version >= '3.8'
pinecone-client==5.0.1; python_version >= '3.8' and python_version < '4.0'
pinecone-plugin-inference==1.0.3; python_version >= '3.8' and python_version < '4.0'
pinecone-plugin-interface==0.0.7; python_version >= '3.8' and python_version < '4.0'
platformdirs==4.2.2; python_version >= '3.8'
pydantic==2.8.2; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
regex==2024.7.24; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
safetensors==0.4.4; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.32; python_version >= '3.7'
tenacity==8.5.0; python_version >= '3.8'
tiktoken==0.7.0; python_version >= '3.8'
tokenizers==0.19.1; python_version >= '3.7'
tqdm==4.66.5; python_version >= '3.7'
transformers==4.44.1; python_full_version >= '3.8.0'
typing-extensions==4.12.2; python_version >= '3.8'
typing-inspect==0.9.0
urllib3==2.2.2; python_version >= '3.12' and python_version < '4.0'
yarl==1.9.4; python_version >= '3.7'
=======
torch
torchvision
onnx
onnxruntime
flask
numpy
streamlit
transformers
scikit-learn
pdfplumber
nltk
PyPDF2
>>>>>>> d0672a9cbff14692ab4e6d9142e08d76b4204046
